+++ ChangeLog.rst
@@ -1,6 +1,9 @@
 Unreleased Changes
 ==================
 
+* There is a new ``timefs3`` example that demonstrates the use of
+  `fuse_lowlevel_notify_inval_entry`.
+
 * There are new ``timefs1`` and ``timefs2`` examples that demonstrate
   the use of the `fuse_lowlevel_notify_store`,
   `fuse_lowlevel_notify_retrieve`, and